requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0

